server/requirements.txt
openenv-core
uvicorn
fastapi
requests
pydantic